+++ src/sections/upload-dataset-files/uploaded-files-list/tag-options-modal/TagOptionsModal.module.scss
